# Improve perf via `isJSON` option