--- a/package.json
+++ b/package.json
@@ -4,7 +4,7 @@
   "description": "Serialize JavaScript to a superset of JSON that includes regular expressions and functions.",
   "main": "index.js",
   "scripts": {
-    "benchmark": "node test/benchmark/serialize.js",
+    "benchmark": "node -v && node test/benchmark/serialize.js",
     "test": "istanbul cover -- ./node_modules/mocha/bin/_mocha test/unit/ --reporter spec"
   },
   "repository": {