@ericf 2016-05-25
Might want to re-think the name of the option given #16 would require the replacer function.